src/test/java/com/jpassion/di/PersonTest.java
package com.jpassion.di;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.SpringApplicationConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

@RunWith(SpringJUnit4ClassRunner.class)
@SpringApplicationConfiguration(classes = MainApplication.class)
public class PersonTest {
	
	@Mock
	private Address address;
	
	@InjectMocks
	private Person person;
	
	@Before
	public void setup() throws Exception {
		MockitoAnnotations.initMocks(this);
	}
	
	@Test
	public void should_return_complete_info_including_address() throws Exception {
		when(address.retrieveAddressInfo()).thenReturn("USA Boston");
		assertThat(person.retrievePersonInfo(), is("Sang Shin USA Boston"));

	}
}
